update(packages): refactor use of ngrx/effects.
Use a StateManagementModule to hide all details of Effects and Reducers.
