+++ .gitignore
@@ -42,3 +42,4 @@ yarn-error.log
 # System Files
 .DS_Store
 Thumbs.db
+/src/environments/environment.ts
