+++ src/app/state/posts/index.ts
@@ -1,0 +1,4 @@
+export * from './post.actions';
+export * from './post.effects';
+export * from './post.model';
+export * from './post.reducer';
