+++ src/app/state/posts/post.effects.ts
@@ -4,13 +4,12 @@ import { AngularFireDatabase }        from 'angularfire2/database';
 
 import { Observable }                 from 'rxjs/Observable';
 import { of }                         from 'rxjs/observable/of';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/mergeMap';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/delay';
 
-import * as postActions from '../actions/post.actions';
-export type Action = postActions.All;
+import '../../utils/rxjs.operators';
+
+import * as postActions from './post.actions';
+
+type Action = postActions.All;
 
 
 @Injectable()
