+++ src/app/utils/rxjs.operators.ts
@@ -1,0 +1,10 @@
+
+import 'rxjs/add/observable/fromPromise';
+import 'rxjs/add/observable/of';
+
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/mergeMap';
+import 'rxjs/add/operator/switchMap';
+
+import 'rxjs/add/operator/catch';
+import 'rxjs/add/operator/delay';
