+++ src/app/views/app.component.scss
